Reorganize code, add setup configuration, and update README
