+++ setup.cfg
@@ -1,0 +1,41 @@
+##########################
+# Setup.py Configuration #
+##########################
+# Configuring setup()
+[metadata]
+name = decimer
+version = 0.0.1-dev
+
+# Author information
+author = Kohulan Rajan
+author_email = kohulan.rajan@uni-jena.de
+maintainer = Kohulan Rajan
+maintainer_email = kohulan.rajan@uni-jena.de
+
+# License information
+license = MIT
+license_file = LICENSE
+
+[options]
+install_requires =
+    tensorflow>=2.3
+    matplotlib
+    pillow
+    selfies
+    unzip
+
+zip_safe = false
+include_package_data = True
+python_requires = >=3.6
+
+# Where is my code
+packages = find:
+package_dir =
+    = src
+
+[options.packages.find]
+where = src
+
+[options.entry_points]
+console_scripts =
+    decimer = decimer.cli:main
