+++ src/decimer/Network/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+"""Network module in the DECIMER Python package."""
