+++ src/decimer/Network/encoder_decoder/TPU_Trainer_Image2Smiles.py
@@ -2,32 +2,34 @@ import tensorflow as tf
 import matplotlib as mpl
 
 mpl.use('Agg')
+
 import matplotlib.pyplot as plt
 
 import re
 import sys
 import time
-import json
 import pickle
-import I2S_Model_Transformer
+import I2S_Model
 from datetime import datetime
-from tensorflow.python.lib.io import file_io
 
-tpu = tf.distribute.cluster_resolver.TPUClusterResolver(tpu='node-blur')
+np.set_printoptions(threshold=sys.maxsize)
+
+tpu = tf.distribute.cluster_resolver.TPUClusterResolver(tpu='node-3')
 print('Running on TPU ', tpu.master())
 
 tf.config.experimental_connect_to_cluster(tpu)
 tf.tpu.experimental.initialize_tpu_system(tpu)
 strategy = tf.distribute.TPUStrategy(tpu)
 
-f = open('Training_Report.txt', 'w')
+f = open('Training_900k1v3-32_128_single_TPU.txt', 'w')
 print('Number of devices: {}'.format(strategy.num_replicas_in_sync), flush=True)
 sys.stdout = f
 print("REPLICAS: ", strategy.num_replicas_in_sync, flush=True)
 print(datetime.now().strftime('%Y/%m/%d %H:%M:%S'), "Network Started", flush=True)
 
 # Load the Data
-total_data =  # Number of total train datasize
+PATH = 'gs://tpu-test-koh/Train_Images/'
+total_data = 999424
 
 tokenizer = pickle.load(open("tokenizer.pkl", "rb"))
 max_length = pickle.load(open("max_length.pkl", "rb"))
@@ -35,26 +37,25 @@ max_length = pickle.load(open("max_length.pkl", "rb"))
 # Parameters to train the network
 
 # Set Training Epochs
-EPOCHS = 60
+EPOCHS = 40
 BATCH_SIZE = 128 * strategy.num_replicas_in_sync
-BUFFER_SIZE = 10000
-target_vocab_size = len(tokenizer.word_index) + 1
+# BATCH_SIZE = 8
+BUFFER_SIZE = 1000
+embedding_dim = 600
+units = 1024
+vocab_size = len(tokenizer.word_index) + 1
 num_steps = total_data // BATCH_SIZE
-num_layer = 4
-d_model = 512
-dff = 2048
-num_heads = 8
-row_size = 10
-col_size = 10
-dropout_rate = 0.1
 
-print("Total train steps: ", num_steps)
+# shape of the vector extracted from InceptionV3 is (64, 2048) these two variables represent that
+features_shape = 2048
+attention_features_shape = 64
 
 AUTO = tf.data.experimental.AUTOTUNE
 
 
 def read_tfrecord(example):
     feature = {
+        # 'image_id': tf.io.FixedLenFeature([], tf.string),
         'image_raw': tf.io.FixedLenFeature([], tf.string),
         'caption': tf.io.FixedLenFeature([], tf.string),
     }
@@ -63,9 +64,8 @@ def read_tfrecord(example):
     example = tf.io.parse_single_example(example, feature)
 
     img = tf.io.decode_raw(example['image_raw'], tf.float32)
-    img_tensor = tf.reshape(img, [100, 1536])
+    img_tensor = tf.reshape(img, [64, 2048])
     caption = tf.io.decode_raw(example['caption'], tf.int32)
-    # caption = tf.reshape(capt, [42,])
 
     return img_tensor, caption
 
@@ -81,8 +81,7 @@ def numericalSort(value):
 
 def get_training_dataset(batch_size=BATCH_SIZE, buffered_size=BUFFER_SIZE):
     options = tf.data.Options()
-    filenames = sorted(tf.io.gfile.glob("path/to/TFRecord/*.tfrecord"), key=numericalSort)
-    # print(len(filenames))
+    filenames = sorted(tf.io.gfile.glob("gs://tpu-test-koh/tfrecords/*.tfrecord"), key=numericalSort)
 
     dataset_img = tf.data.TFRecordDataset(filenames, num_parallel_reads=AUTO)
 
@@ -97,43 +96,12 @@ def get_training_dataset(batch_size=BATCH_SIZE, buffered_size=BUFFER_SIZE):
     return train_dataset
 
 
-class CustomSchedule(tf.keras.optimizers.schedules.LearningRateSchedule):
-    def __init__(self, d_model, warmup_steps=4000):
-        super(CustomSchedule, self).__init__()
-
-        self.d_model = d_model
-        self.d_model = tf.cast(self.d_model, tf.float32)
-
-        self.warmup_steps = warmup_steps
-
-    def __call__(self, step):
-        arg1 = tf.math.rsqrt(step)
-        arg2 = step * (self.warmup_steps ** -1.5)
-
-        return tf.math.rsqrt(self.d_model) * tf.math.minimum(arg1, arg2)
-
-
-def create_padding_mask(seq):
-    seq = tf.cast(tf.math.equal(seq, 0), tf.float32)
-    return seq[:, tf.newaxis, tf.newaxis, :]
-
-
-def create_look_ahead_mask(size):
-    mask = 1 - tf.linalg.band_part(tf.ones((size, size)), -1, 0)
-    return mask
-
-
-def create_masks_decoder(tar):
-    look_ahead_mask = create_look_ahead_mask(tf.shape(tar)[1])
-    dec_target_padding_mask = create_padding_mask(tar)
-    combined_mask = tf.maximum(dec_target_padding_mask, look_ahead_mask)
-    return combined_mask
-
-
 with strategy.scope():
+    encoder = I2S_Model.CNN_Encoder(embedding_dim)
+    decoder = I2S_Model.RNN_Decoder(embedding_dim, units, vocab_size)
+
     # Network Parameters
-    learning_rate = CustomSchedule(d_model)
-    optimizer = tf.keras.optimizers.Adam(learning_rate, beta_1=0.9, beta_2=0.98, epsilon=1e-9)
+    optimizer = tf.keras.optimizers.Adam(learning_rate=0.00051)
     loss_object = tf.keras.losses.SparseCategoricalCrossentropy(from_logits=True, reduction='none')
 
 
@@ -146,17 +114,8 @@ with strategy.scope():
 
         return tf.reduce_mean(loss_)
 
-
-    train_loss = tf.keras.metrics.Mean(name='train_loss', dtype=tf.float32)
-    train_accuracy = tf.keras.metrics.SparseCategoricalAccuracy(name='train_accuracy', dtype=tf.float32)
-
-    # Initialize Transformer
-    transformer = I2S_Model_Transformer.Transformer(num_layer, d_model, num_heads, dff, row_size, col_size,
-                                                    target_vocab_size, max_pos_encoding=target_vocab_size,
-                                                    rate=dropout_rate)
-
-checkpoint_path = "path/to/checkpoint"
-ckpt = tf.train.Checkpoint(transformer=transformer, optimizer=optimizer)
+checkpoint_path = "gs://tpu-test-koh/checkpoints/train_900k1v3-32_128_TPU_Test_pred"
+ckpt = tf.train.Checkpoint(encoder=encoder, decoder=decoder, optimizer=optimizer)
 ckpt_manager = tf.train.CheckpointManager(ckpt, checkpoint_path, max_to_keep=50)
 
 start_epoch = 0
@@ -172,7 +131,6 @@ train_dataset = strategy.experimental_distribute_dataset(get_training_dataset())
 
 # the loss_plot array will be reset many times
 loss_plot = []
-accuracy_plot = []
 
 
 @tf.function
@@ -181,24 +139,30 @@ def train_step(iterator):
         img_tensor, target = inputs
         loss = 0
 
-        tar_inp = target[:, :-1]
-        tar_real = target[:, 1:]
+        # initializing the hidden state for each batch because the captions are not related from image to image
+        hidden = decoder.reset_state(batch_size=target.shape[0])
 
-        dec_mask = create_masks_decoder(tar_inp)
-        print(dec_mask)
+        dec_input = tf.expand_dims([tokenizer.word_index['<start>']] * target.shape[0], 1)
 
         with tf.GradientTape() as tape:
-            predictions, _ = transformer(img_tensor, tar_inp, True, dec_mask)
-            loss = loss_function(tar_real, predictions)
+            features = encoder(img_tensor)
+
+            for i in range(1, target.shape[1]):
+                # passing the features through the decoder
+                predictions, hidden, _ = decoder(dec_input, features, hidden)
+
+                loss += loss_function(target[:, i], predictions)
+
+                # using teacher forcing
+                dec_input = tf.expand_dims(target[:, i], 1)
 
         total_loss = (loss / int(target.shape[1]))
 
-        gradients = tape.gradient(loss, transformer.trainable_variables)
+        trainable_variables = encoder.trainable_variables + decoder.trainable_variables
 
-        optimizer.apply_gradients(zip(gradients, transformer.trainable_variables))
+        gradients = tape.gradient(loss, trainable_variables)
 
-        train_loss.update_state(loss * strategy.num_replicas_in_sync)
-        train_accuracy.update_state(tar_real, predictions)
+        optimizer.apply_gradients(zip(gradients, trainable_variables))
 
         return loss, total_loss
 
@@ -215,9 +179,6 @@ for epoch in range(start_epoch, EPOCHS):
     start = time.time()
     total_loss = 0
     batch = 0
-    train_loss.reset_states()
-    train_accuracy.reset_states()
-    print(str(learning_rate))
 
     for x in train_dataset:
         img_tensor, target = x
@@ -226,24 +187,15 @@ for epoch in range(start_epoch, EPOCHS):
         batch += 1
 
         if batch % 100 == 0:
-            print('Epoch {} Batch {} Loss {:.4f} Updated_Loss {:.4f} Accuracy {:.4f}'.format(epoch + 1, batch,
-                                                                                             batch_loss.numpy() / BATCH_SIZE,
-                                                                                             train_loss.result(),
-                                                                                             train_accuracy.result()),
-                  flush=True)
-        # storing the epoch end loss value to plot later
+            print('Epoch {} Batch {} Loss {:.4f}'.format(epoch + 1, batch, batch_loss.numpy() / BATCH_SIZE), flush=True)
 
         if batch == num_steps:
             loss_plot.append(total_loss / num_steps)
             ckpt_manager.save()
 
-            print('Epoch {} Loss {:.6f} Updated_Loss {:.4f} Accuracy {:.4f}'.format(epoch + 1, total_loss / num_steps,
-                                                                                    train_loss.result(),
-                                                                                    train_accuracy.result()),
-                  flush=True)
+            print('Epoch {} Loss {:.6f}'.format(epoch + 1, total_loss / num_steps), flush=True)
             print(datetime.now().strftime('%Y/%m/%d %H:%M:%S'),
                   'Time taken for 1 epoch {} sec\n'.format(time.time() - start), flush=True)
-            # transformer.save_weights('Epoch_'+str(epoch+1)+'_weights.h5')
 
             break
 
@@ -253,7 +205,7 @@ plt.xlabel('Epochs')
 plt.ylabel('Loss')
 # plt.show()
 plt.gcf().set_size_inches(20, 20)
-plt.savefig("Lossplot_.jpg")
+plt.savefig("Lossplot_900k1v3-32_128_tpu_test.jpg")
 plt.close()
 
 print(datetime.now().strftime('%Y/%m/%d %H:%M:%S'), "Network Completed", flush=True)
