+++ src/decimer/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+"""DECIMER Python Package."""
