+++ src/decimer/__main__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+"""DECIMER CLI."""
+
+from .cli import main
+
+if __name__ == '__main__':
+    main()
